// src/app/services/interfaces/api.interfaces.ts
export interface OFund {
    tipo: string;
    nfund: NFund[];
  }
  
export interface NFund {
    nombre: string;
    fund: Fundamento[];
  }
// ...
export interface Fundamento {
    tramiteId:          number;
    tipo:               string;
    requisitoId:        number | null;
    canalAtencionId:    number | null;
    formatoId:          number | null;
    plazoPrevencionId:  null;
    plazoInterezadoId:  null;
    plazoMaximoId:      number | null;
    ordenGobierno:      OrdenGobierno;
    tipoOrdenamiento:   TipoOrdenamiento;
    nombreOrdenamiento: NombreOrdenamiento;
    articulo:           string;
    fraccion:           null | string;
    inciso:             null | string;
    parrafo:            null | string;
    numero:             null | string;
    letra:              null | string;
    otro:               null | string;
}

export enum NombreOrdenamiento {
    CódigoCivilParaElEstadoDeSANLuisPotosí = "Código Civil para el Estado de San Luis Potosí",
    LeyDeHaciendaParaElEstadoDeSANLuisPotosi = "Ley de Hacienda para el Estado de San Luis Potosi",
    LeyDelRegistroCivilDelEstadoDeSANLuisPotosí = "Ley del Registro Civil del Estado de San Luis Potosí",
    ReglamentoDeLaLeyDelRegistroCivilDelEstado = "Reglamento de la Ley del Registro Civil del Estado",
    ReglamentoInteriorDeLaSecretaríaGeneralDeGobierno = "Reglamento Interior de la Secretaría General de Gobierno",
}

export enum OrdenGobierno {
    Estatal = "Estatal",
}

export enum TipoOrdenamiento {
    Código = "Código",
    Ley = "Ley",
    Reglamento = "Reglamento",
}